# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This api is for testing the swagger ui",
    "version": "1.0.7",
    "title": "Express API",
    "contact": {
      "email": "test@gmail.com"
    }
  },
  "host": "0.0.0.0:3000",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/{userId}": {
      "get": {
        "tags": ["user"],
        "summary": "Get user by ID",
        "description": "Retrieve details of a user by their unique ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/PrivateUserResponse"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/users/me/avatar": {
      "post": {
        "tags": ["user"],
        "summary": "Upload avatar",
        "description": "Uploads an avatar for the user",
        "operationId": "uploadAvatar",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "avatar",
            "in": "formData",
            "description": "Avatar file to upload",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar uploaded successfully"
          },
          "400": {
            "description": "Invalid file format"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "summary": "Delete avatar",
        "description": "Deletes the user's avatar",
        "operationId": "deleteAvatar",
        "responses": {
          "204": {
            "description": "Avatar deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "definitions": {
    "PrivateUserResponse": {
      "type": "object",
      "required": ["_id", "name", "email", "role", "isDeleted", "isVerified"],
      "properties": {
        "_id": {
          "type": "string",
          "format": "uuid",
          "description": "The unique identifier for a user"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "age": {
          "type": "number"
        },
        "phone": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": ["admin", "user"]
        },
        "isDeleted": {
          "type": "boolean"
        },
        "isVerified": {
          "type": "boolean"
        }
      }
    },
...
  }
}